cagey/baogang/requirements.txt
Package                  Version  
------------------------ ---------
asn1crypto               0.24.0   
attrs                    19.1.0   
Automat                  0.7.0    
beautifulsoup4           4.8.0    
bitarray                 1.0.1    
boltons                  19.1.0   
certifi                  2019.6.16
cffi                     1.12.3   
chardet                  3.0.4    
Click                    7.0      
constantly               15.1.0   
cryptography             2.7      
cssselect                1.1.0    
demjson                  2.2.4    
Django                   2.2.4    
django-cors-headers      3.1.0    
docker                   4.1.0    
docker-py                1.10.6   
docker-pycreds           0.4.0    
docopt                   0.6.2    
face                     19.1.2   
Flask                    1.1.1    
fonttools                4.0.2    
gerapy                   0.8.5    
gevent                   1.4.0    
glom                     19.2.0   
greenlet                 0.4.15   
html5lib                 1.0.1    
HTMLParser               0.0.2    
hyperlink                19.0.0   
idna                     2.8      
incremental              17.5.0   
itsdangerous             1.1.0    
Jinja2                   2.10.1   
lxml                     4.4.1    
MarkupSafe               1.1.1    
mysqlclient              1.4.4    
numpy                    1.17.1   
pandas                   0.25.1   
parsel                   1.5.2    
Pillow                   6.1.0    
pip                      19.3.1   
psutil                   5.6.4    
pyasn1                   0.4.6    
pyasn1-modules           0.2.6    
pybloom-live             3.0.0    
pycparser                2.19     
PyDispatcher             2.0.5    
PyExecJS                 1.5.1    
PyHamcrest               1.9.0    
pymongo                  3.9.0    
PyMySQL                  0.9.3    
pyOpenSSL                19.0.0   
pytesseract              0.3.0    
python-dateutil          2.8.0    
python-scrapyd-api       2.1.2    
pytidylib                0.3.2    
pytz                     2019.2   
queuelib                 1.5.0    
redis                    3.3.8    
requests                 2.22.0   
Scrapy                   1.7.3    
scrapy-redis             0.6.8    
scrapy-redis-bloomfilter 0.7.0    
scrapy-splash            0.7.2    
selenium                 3.141.0  
service-identity         18.1.0   
setuptools               40.8.0   
six                      1.12.0   
soupsieve                1.9.4    
SQLAlchemy               1.3.8    
sqlparse                 0.3.0    
treelib                  1.5.5    
Twisted                  19.7.0   
urllib3                  1.25.3   
w3lib                    1.21.0   
web.py                   0.40.dev0
webencodings             0.5.1    
webpage                  0.1      
websocket-client         0.56.0   
Werkzeug                 0.16.0   
xlrd                     1.2.0    
zope.interface           4.6.0    
